Refactor away duplicated *-tsc files
