+++ package.json
@@ -2,24 +2,22 @@
   "name": "@mishguru/package",
   "version": "6.0.0-next.10",
   "description": "Utils for creating, building and testing NPM packages",
-  "main": "index.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "build": "rm -rf ./dist && tsc",
-    "test": "./bin/test-tsc",
-    "lint": "./bin/lint-tsc",
+    "test": "./bin/test",
+    "lint": "./bin/lint",
     "tidy": "./bin/tidy"
   },
   "bin": {
     "pkg-precommit": "./bin/precommit",
     "pkg-build": "./bin/build",
-    "pkg-build-tsc": "./bin/build-tsc",
     "pkg-clean": "./bin/clean",
     "pkg-coverage": "./bin/coverage",
     "pkg-flow": "./bin/flow",
     "pkg-lint": "./bin/lint",
-    "pkg-lint-tsc": "./bin/lint-tsc",
     "pkg-test": "./bin/test",
-    "pkg-test-tsc": "./bin/test-tsc",
     "pkg-tidy": "./bin/tidy"
   },
   "repository": {
