+++ src/cmd/build.ts
@@ -1,23 +1,29 @@
-import { SRC_PATH, DIST_PATH } from '../shared/constants'
+import { USE_TSC, SRC_PATH, DIST_PATH } from '../shared/constants'
 import { log, fmt } from '../shared/log'
-
-import babel from '../babel'
-import BABEL_CONFIG from '../babel/config'
+import exec from '../shared/exec'
 
 import clean from '../cmd/clean'
 
-const flowCopy = require('flow-copy-source')
+import babel from '../babel'
+import BABEL_CONFIG from '../babel/config'
+import flowCopySource from 'flow-copy-source'
+
+const TSC_SHIM_PATH = require.resolve('../shim/tsc')
 
 const build = async () => {
   await clean()
 
   log(fmt`Building from ${SRC_PATH} to ${DIST_PATH} directory`)
-  await babel(SRC_PATH, DIST_PATH, BABEL_CONFIG)
 
-  log(
-    fmt`Copying flow source files from ${SRC_PATH} to ${DIST_PATH} directory...`,
-  )
-  await flowCopy([SRC_PATH], DIST_PATH)
+  if (USE_TSC) {
+    const args = process.argv.slice(2)
+    log(fmt`Running ${'tsc'} ${args}`)
+    await exec('node', TSC_SHIM_PATH, ...args)
+  } else {
+    await babel(SRC_PATH, DIST_PATH, BABEL_CONFIG)
+    log(fmt`Copying flow source files...`)
+    await flowCopySource([SRC_PATH], DIST_PATH)
+  }
 }
 
 export default build
