+++ src/cmd/lint.ts
@@ -1,9 +1,13 @@
 import execLinter from '../shared/execLinter'
+import { USE_TSC } from '../shared/constants'
 
-const LINTER_PATH = require.resolve('../shim/eslint')
+const LINTER_PATH = USE_TSC
+  ? require.resolve('../shim/eslint-tsc')
+  : require.resolve('../shim/eslint')
 
 const lint = async () => {
-  await execLinter(LINTER_PATH)
+  const args = process.argv.slice(2)
+  await execLinter(LINTER_PATH, args)
 }
 
 export default lint
