+++ src/cmd/test.ts
@@ -1,8 +1,11 @@
 import withTestHelpers from '../shared/withTestHelpers'
 import exec from '../shared/exec'
 import { log, fmt } from '../shared/log'
+import { USE_TSC } from '../shared/constants'
 
-const AVA_PATH = require.resolve('../shim/ava')
+const AVA_PATH = USE_TSC
+  ? require.resolve('../shim/ava-tsc')
+  : require.resolve('../shim/ava')
 
 const test = async () => {
   await withTestHelpers(async () => {
