+++ src/cmd/tidy.ts
@@ -1,17 +1,14 @@
-import { SRC_PATH } from '../shared/constants'
-
+import { USE_TSC, SRC_PATH } from '../shared/constants'
 import exec from '../shared/exec'
 import { log, fmt } from '../shared/log'
+import execLinter from '../shared/execLinter'
+
+import config from '../config/prettier'
 
 const PRETTIER_PATH = require.resolve('prettier/bin-prettier')
-
-const config = [
-  '--single-quote',
-  '--trailing-comma=all',
-  '--arrow-parens=always',
-  '--no-semi',
-  '--write',
-]
+const LINTER_PATH = USE_TSC
+  ? require.resolve('../shim/eslint-tsc')
+  : require.resolve('../shim/eslint')
 
 const prettier = async () => {
   const args = process.argv.slice(2)
@@ -23,6 +20,8 @@ const prettier = async () => {
       : [`./${SRC_PATH}/**/*.{js,jsx,css,scss,ts,md,yml,json,less}`]
 
   await exec('node', PRETTIER_PATH, ...config, ...files)
+
+  await execLinter(LINTER_PATH, ['--fix'])
 }
 
 export default prettier
