+++ src/config/prettier.ts
@@ -1,0 +1,9 @@
+const config = [
+  '--single-quote',
+  '--trailing-comma=all',
+  '--arrow-parens=always',
+  '--no-semi',
+  '--write',
+]
+
+export default config
