+++ src/shared/execLinter.ts
@@ -2,15 +2,14 @@ import { SRC_PATH } from '../shared/constants'
 import exec from '../shared/exec'
 import { log, fmt } from '../shared/log'
 
-const execLinter = async (linterPath: string) => {
-  const args = process.argv.slice(2)
+const execLinter = async (linterPath: string, args: string[]) => {
   log(fmt`Running ${'eslint'} ${args}`)
 
   const options = args.filter((input) => input.startsWith('--'))
   if (options.length > 0) {
-    const fixIndex = args.indexOf(options[0])
-    if (fixIndex > -1) {
-      args.splice(fixIndex, options.length)
+    const index = args.indexOf(options[0])
+    if (index > -1) {
+      args.splice(index, options.length)
     }
   }
