+++ src/types/flow-copy-source.d.ts
@@ -1,0 +1,3 @@
+declare module 'flow-copy-source' {
+  export default function(source: string[], dest: string): Promise<void>
+}
